K to int

+++ train_distributed.py
@@ -142,7 +142,7 @@ def train(opt):
     try:
         ### 수정된 부분 ###
         num_of_workers = int(os.environ["WORLD_SIZE"])
-        K = np.ceil((epoch + 1)**(1/3) / num_of_workers)
+        K = int(np.ceil((epoch + 1)**(1/3) / num_of_workers))
         counter = 0
         ###################
 
@@ -188,7 +188,7 @@ def train(opt):
                 ### 수정된 부분 ###
                 # exchange
                 average_weights(model)
-                K = np.ceil((epoch + 1)**(1/3) / num_of_workers)
+                K = int(np.ceil((epoch + 1)**(1/3) / num_of_workers))
                 ###################
